A few typecasts to enable / simplify 64-bit compilation
git-svn-id: svn://ultimatepp.org/upp/trunk@1653 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Algo.h
@@ -1000,7 +1000,7 @@ template <class II, class VI, class K, class Less>
 inline void __IndexSort(II begin, II end, VI pair, const Less& less, const K *)
 {
 	Sort(IndexSortIterator<II, VI, K>(begin, pair),
-		IndexSortIterator<II, VI, K>(end, pair + (end - begin)),
+		IndexSortIterator<II, VI, K>(end, pair + (int)(ptrdiff_t)(end - begin)),
 		less);
 }
